+++ Assets/VRExplorer/CHANGELOG.md
@@ -3,5 +3,6 @@
 ## [1.5.5] - 2025-06-18
 ### Added
 - Support for exporting GameObjects with scripts under the `VRExplorer` namespace only.
-- Stable identifier logic using `GlobalObjectId`（scene objects）& AssetDatabase GUID（prefabs）
+- Stable identifier logic using `GlobalObjectId` (scene objects) & AssetDatabase GUID(prefabs).
 
+- `RemoveVRExplorerScripts()` for remove all the added VRExplorer Mono predefined scripts.